refactor(ui): split shell chrome modules
Closes #2639
