+++ apps/ui/src/app/runtime/shell/ui_shell_chrome_hosts.tsx
@@ -1,0 +1,144 @@
+import type { ComponentChildren } from "preact";
+import { useRef } from "preact/hooks";
+
+import {
+  useComputed,
+  useSignalProperties,
+  type ReadonlySignal,
+} from "../../ui_signals";
+import {
+  SHELL_ACTIVE_VIEW_KEY,
+  type UiShellChromeNavigationModel,
+  type UiShellChromePendingPanelHosts,
+} from "./ui_shell_chrome_shared";
+
+type ShellViewSectionProps = {
+  activeViewId: ReadonlySignal<string>;
+  ariaLabelledBy: string;
+  children: ComponentChildren;
+  viewId: string;
+};
+
+export function ShellViewHostsContainer(props: {
+  navigationModel: ReadonlySignal<UiShellChromeNavigationModel>;
+  panelHosts: UiShellChromePendingPanelHosts;
+}) {
+  const { navigationModel, panelHosts } = props;
+  const { activeViewId } = useSignalProperties(navigationModel, SHELL_ACTIVE_VIEW_KEY);
+
+  return (
+    <>
+      <ShellViewSection
+        activeViewId={activeViewId}
+        ariaLabelledBy="tab-dashboard"
+        viewId="dashboardView"
+      >
+        <DashboardViewHosts panelHosts={panelHosts.dashboard} />
+      </ShellViewSection>
+
+      <ShellViewSection
+        activeViewId={activeViewId}
+        ariaLabelledBy="tab-history"
+        viewId="historyView"
+      >
+        <HistoryViewHosts panelHosts={panelHosts} />
+      </ShellViewSection>
+
+      <ShellViewSection
+        activeViewId={activeViewId}
+        ariaLabelledBy="tab-settings"
+        viewId="settingsView"
+      >
+        <SettingsViewHosts panelHosts={panelHosts} />
+      </ShellViewSection>
+    </>
+  );
+}
+
+function DashboardViewHosts(props: {
+  panelHosts: UiShellChromePendingPanelHosts["dashboard"];
+}) {
+  const { panelHosts } = props;
+  const liveOverviewHostRef = useRef<HTMLDivElement | null>(null);
+  const spectrumHostRef = useRef<HTMLDivElement | null>(null);
+  const loggingHostRef = useRef<HTMLDivElement | null>(null);
+
+  return (
+    <div class="dashboard-grid">
+      <div
+        id="liveOverviewRoot"
+        ref={(element) => {
+          liveOverviewHostRef.current = element;
+          panelHosts.liveOverview = element;
+        }}
+        class="panel card dashboard-grid__overview"
+      ></div>
+      <div
+        id="spectrumPanelRoot"
+        ref={(element) => {
+          spectrumHostRef.current = element;
+          panelHosts.spectrum = element;
+        }}
+        class="panel card dashboard-grid__main"
+      ></div>
+      <div
+        id="loggingPanelRoot"
+        ref={(element) => {
+          loggingHostRef.current = element;
+          panelHosts.logging = element;
+        }}
+        class="panel card dashboard-grid__controls"
+      ></div>
+    </div>
+  );
+}
+
+function HistoryViewHosts(props: {
+  panelHosts: UiShellChromePendingPanelHosts;
+}) {
+  const hostRef = useRef<HTMLDivElement | null>(null);
+
+  return (
+    <div
+      id="historyPanelRoot"
+      ref={(element) => {
+        hostRef.current = element;
+        props.panelHosts.history = element;
+      }}
+      class="panel card"
+    ></div>
+  );
+}
+
+function SettingsViewHosts(props: {
+  panelHosts: UiShellChromePendingPanelHosts;
+}) {
+  const hostRef = useRef<HTMLDivElement | null>(null);
+
+  return (
+    <div
+      id="settingsShellRoot"
+      ref={(element) => {
+        hostRef.current = element;
+        props.panelHosts.settingsShell = element;
+      }}
+    ></div>
+  );
+}
+
+function ShellViewSection(props: ShellViewSectionProps) {
+  const { activeViewId, ariaLabelledBy, children, viewId } = props;
+  const hidden = useComputed(() => activeViewId.value !== viewId);
+
+  return (
+    <section
+      id={viewId}
+      class="view"
+      role="tabpanel"
+      aria-labelledby={ariaLabelledBy}
+      hidden={hidden}
+    >
+      {children}
+    </section>
+  );
+}
