+++ apps/ui/src/app/runtime/shell/ui_shell_chrome_status.tsx
@@ -1,0 +1,66 @@
+import type { ComponentChildren } from "preact";
+
+import {
+  useComputed,
+  useSignalProperties,
+  type ReadonlySignal,
+} from "../../ui_signals";
+import {
+  SHELL_ACTIVE_VIEW_KEY,
+  SHELL_STATUS_MODEL_KEYS,
+  type UiShellBadgeModel,
+  type UiShellChromeNavigationModel,
+  type UiShellChromeStatusModel,
+} from "./ui_shell_chrome_shared";
+
+export function ShellChromeFrame(props: {
+  children: ComponentChildren;
+  statusModel: ReadonlySignal<UiShellChromeStatusModel>;
+}) {
+  const { children, statusModel } = props;
+  const connectionState = useComputed(() => statusModel.value.connectionState);
+
+  return (
+    <div class="wrap" data-connection-state={connectionState}>
+      {children}
+    </div>
+  );
+}
+
+export function ShellStatus(props: {
+  navigationModel: ReadonlySignal<UiShellChromeNavigationModel>;
+  statusModel: ReadonlySignal<UiShellChromeStatusModel>;
+}) {
+  const { navigationModel, statusModel } = props;
+  const { activeViewId } = useSignalProperties(navigationModel, SHELL_ACTIVE_VIEW_KEY);
+  const { shellLiveStatus, wsLinkState } = useSignalProperties(statusModel, SHELL_STATUS_MODEL_KEYS);
+  const statusHidden = useComputed(() => activeViewId.value === "dashboardView");
+
+  return (
+    <div class="site-header__status" hidden={statusHidden}>
+      <div class="site-header__status-pills">
+        <ShellStatusPill id="linkState" model={wsLinkState} />
+        <ShellStatusPill id="shellLiveStatus" model={shellLiveStatus} />
+      </div>
+    </div>
+  );
+}
+
+function ShellStatusPill(props: {
+  id: string;
+  model: ReadonlySignal<UiShellBadgeModel>;
+}) {
+  const variant = useComputed(() => props.model.value.variant);
+  const text = useComputed(() => props.model.value.text);
+
+  return (
+    <div
+      id={props.id}
+      class="pill"
+      data-variant={variant}
+      aria-live="polite"
+    >
+      {text}
+    </div>
+  );
+}
